+++ src/Persistence/src/Service/AircraftService.cpp
@@ -73,13 +73,11 @@ bool AircraftService::store(std::int64_t flightId, std::size_t sequenceNumber, A
             ok = d->flightDao->updateUserAircraftIndex(flight.getId(), flight.getUserAircraftIndex());
         }
         if (ok) {
-            ok = db.commit();
-            if (ok) {
-                emit flight.aircraftStored();
-            }
+            ok = db.commit();            
         } else {
             db.rollback();
         }
+        emit flight.aircraftStored(ok);
     }
     return ok;
 }
@@ -96,12 +94,10 @@ bool AircraftService::exportAircraft(std::int64_t flightId, std::size_t sequence
         }
         if (ok) {
             ok = db.commit();
-            if (ok) {
-                emit flight.aircraftStored();
-            }
         } else {
             db.rollback();
         }
+        emit flight.aircraftStored(ok);
     }
     return ok;
 }