+++ src/Plugins/Module/Formation/src/FormationWidget.cpp
@@ -276,6 +276,8 @@ void FormationWidget::frenchConnection() noexcept
     Flight &flight = Logbook::getInstance().getCurrentFlight();
     connect(&flight, &Flight::flightRestored,
             this, &FormationWidget::updateUi);
+    connect(&flight, &Flight::aircraftStored,
+            this, &FormationWidget::updateUi);
     connect(&flight, &Flight::userAircraftChanged,
             this, &FormationWidget::onUserAircraftChanged);
     connect(&flight, &Flight::aircraftAdded,