+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Flight condition is not overwritten anymore when recording additional formation aircraft
 - Toggling the "active pause" in MSFS while recording now properly resumes/pauses recording and the flight itself [[Issue #106](https://github.com/till213/SkyDolly/issues/106)]
 - Prevent a crash that would occur when a flight is loaded and the replay mode in the Formation module is changed to "Fly with formation", but no connection with MSFS exists yet
+- Remove last recorded flight or aircraft in case persistence fails (improved error handling)
 
 ### Under the Hood
 - Upgrade GeographicLib to version 2.2 (from version 2.1.2)